refactor(search): minor constants-use and types refactoring in domain abstractions

+++ planners-search/src/numeric/evaluation/domain_abstractions/cegar.rs
@@ -211,7 +211,7 @@ impl Cegar {
         let state_packer = make_prop_state_packer(task);
         let axiom_evaluator = AxiomEvaluator::new(task, &state_packer);
 
-        let mut buffer = vec![0u64; state_packer.num_bins() as usize];
+        let mut buffer = vec![0u64; state_packer.num_bins()];
         set_initial_prop_values(task, &state_packer, &mut buffer);
         let mut numeric_state: Vec<f64> = task.get_initial_numeric_state_values().to_vec();
